Move AstPragmaTransformer into AstTransforms
